requirements.txt
asgiref==3.5.2
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==2.1.0
cloudinary==1.30.0
confusable-homoglyphs==3.2.0
cryptography==37.0.4
dj-database-url==0.5.0
Django==4.1.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-on-heroku==1.1.2
django-registration==3.3
djangorestframework==3.14.0
djangorestframework-pagination==1.1
django-tinymce==3.4.0
django-cors-headers==3.13.0
idna==3.4                                                                                                                                                                          
psycopg2-binary==2.9.9
pycodestyle==2.7.0
pycparser==2.21
python-decouple==3.5
pytz==2022.6
requests==2.28.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
whitenoise==6.2.0
zipp==3.6.0
gunicorn==20.1.0
pyuploadcare==3.1.0
django-simple-mail==2.3.5
